[Fluid Clean]Migrate utils files and delete dygraph_to_static dir
