+++ python/paddle/fluid/dygraph/__init__.py
@@ -43,8 +43,6 @@ from .io import *
 from . import static_runner
 from .static_runner import StaticModelRunner
 
-from . import dygraph_to_static
-
 from . import rnn
 from .rnn import *
