--- python/paddle/fluid/dygraph/dygraph_to_static/__init__.py
@@ -1,28 +1,0 @@
-#   Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from . import static_analysis
-from .static_analysis import *
-
-from . import variable_trans_func
-from .variable_trans_func import *
-
-
-from . import logging_utils
-from .logging_utils import *
-
-__all__ = []
-__all__ += static_analysis.__all__
-__all__ += variable_trans_func.__all__
-__all__ += logging_utils.__all__
