+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_basic_api_transformation.py
@@ -22,7 +22,7 @@ import paddle.fluid as fluid
 import paddle.fluid.dygraph as dygraph
 from paddle import to_tensor
 from paddle.fluid.dygraph import to_variable
-from paddle.fluid.dygraph.dygraph_to_static.utils import is_dygraph_api
+from paddle.jit.dy2static.utils import is_dygraph_api
 from paddle.jit.api import dygraph_to_static_func
 from paddle.utils import gast
