+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_break_continue.py
@@ -18,7 +18,7 @@ import numpy as np
 
 import paddle
 import paddle.fluid as fluid
-from paddle.fluid.dygraph.dygraph_to_static.utils import Dygraph2StaticException
+from paddle.jit.dy2static.utils import Dygraph2StaticException
 from paddle.jit.api import declarative
 from paddle.jit.dy2static.program_translator import ProgramTranslator
