+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_closure_analysis.py
@@ -18,7 +18,7 @@ import unittest
 from numpy import append
 
 import paddle
-from paddle.fluid.dygraph.dygraph_to_static.utils import (
+from paddle.jit.dy2static.utils import (
     FunctionNameLivenessAnalysis,
 )
 from paddle.utils import gast
