+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_ifelse.py
@@ -43,7 +43,7 @@ from ifelse_simple_func import (
 
 import paddle
 import paddle.fluid.core as core
-from paddle.fluid.dygraph.dygraph_to_static.utils import Dygraph2StaticException
+from paddle.jit.dy2static.utils import Dygraph2StaticException
 from paddle.jit.api import declarative
 from paddle.jit.dy2static.program_translator import ProgramTranslator
