+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_logging_utils.py
@@ -20,7 +20,7 @@ import unittest
 from unittest import mock
 
 import paddle
-from paddle.fluid.dygraph.dygraph_to_static import logging_utils
+from paddle.jit.dy2static import logging_utils
 from paddle.utils import gast
