+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_origin_info.py
@@ -15,7 +15,7 @@
 import sys
 import unittest
 
-from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
+from paddle.jit.dy2static.origin_info import (
     ORIGI_INFO,
     Location,
     OriginInfo,
@@ -25,7 +25,7 @@ from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
     inspect,
     unwrap,
 )
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_func
+from paddle.jit.dy2static.utils import ast_to_func
 from paddle.jit.api import declarative
 from paddle.jit.dy2static import DygraphToStaticAst
