+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_rollback.py
@@ -17,7 +17,7 @@ import unittest
 import numpy as np
 
 import paddle
-from paddle.fluid.dygraph.dygraph_to_static.utils import func_to_source_code
+from paddle.jit.dy2static.utils import func_to_source_code
 from paddle.jit.dy2static.program_translator import StaticFunction
