+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_variable_trans_func.py
@@ -14,8 +14,8 @@
 
 import unittest
 
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
-from paddle.fluid.dygraph.dygraph_to_static.variable_trans_func import (
+from paddle.jit.dy2static.utils import ast_to_source_code
+from paddle.jit.dy2static.variable_trans_func import (
     create_fill_constant_node,
 )
