+++ python/paddle/fluid/tests/unittests/test_input_spec.py
@@ -20,7 +20,7 @@ import numpy as np
 
 import paddle
 import paddle.fluid as fluid
-from paddle.fluid.dygraph.dygraph_to_static.utils import (
+from paddle.jit.dy2static.utils import (
     _compatible_non_tensor_spec,
 )
 from paddle.fluid.framework import convert_np_dtype_to_dtype_
