+++ python/paddle/jit/api.py
@@ -34,16 +34,16 @@ from paddle.fluid.dygraph.base import (
     program_desc_tracing_guard,
     switch_to_static_graph,
 )
-from paddle.fluid.dygraph.dygraph_to_static import logging_utils
-from paddle.jit.dy2static.convert_call_func import (
+from . import logging_utils
+from .convert_call_func import (
     ConversionOptions,
     CONVERSION_OPTIONS,
 )
-from paddle.fluid.dygraph.dygraph_to_static.logging_utils import (
+from .logging_utils import (
     set_code_level,
     set_verbosity,
 )
-from paddle.jit.dy2static.program_translator import (
+from .program_translator import (
     ProgramTranslator,
     StaticFunction,
     unwrap_decorators,
