+++ python/paddle/jit/dy2static/basic_api_transformer.py
@@ -15,14 +15,13 @@
 import astor
 from paddle.utils import gast
 
-from paddle.fluid.dygraph.dygraph_to_static.static_analysis import (
+from .static_analysis import (
     AstNodeWrapper,
 )
-from paddle.fluid.dygraph.dygraph_to_static import utils
+from . import utils
 from .base_transformer import (
     BaseTransformer,
 )
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
 
 
 class BasicApiTransformer(BaseTransformer):
@@ -166,7 +165,7 @@ class AttributeJstTransformer(BaseTransformer):
             node = (
                 gast.parse(
                     "_jst.Attr({}, \"{}\")".format(
-                        ast_to_source_code(value).strip(), attr
+                        utils.ast_to_source_code(value).strip(), attr
                     )
                 )
                 .body[0]
