+++ python/paddle/jit/dy2static/break_continue_transformer.py
@@ -15,9 +15,9 @@
 from paddle.utils import gast
 
 from paddle.fluid import unique_name
-from paddle.fluid.dygraph.dygraph_to_static.utils import index_in_list
-from paddle.fluid.dygraph.dygraph_to_static.utils import BaseNodeVisitor
-from paddle.fluid.dygraph.dygraph_to_static.variable_trans_func import (
+from paddle.jit.dy2static.utils import index_in_list
+from paddle.jit.dy2static.utils import BaseNodeVisitor
+from paddle.jit.dy2static.variable_trans_func import (
     create_bool_node,
 )
 from .base_transformer import (
