+++ python/paddle/jit/dy2static/call_transformer.py
@@ -14,11 +14,11 @@
 
 from paddle.utils import gast
 
-from paddle.fluid.dygraph.dygraph_to_static.static_analysis import (
+from paddle.jit.dy2static.static_analysis import (
     AstNodeWrapper,
 )
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
-from paddle.fluid.dygraph.dygraph_to_static.utils import is_paddle_api
+from paddle.jit.dy2static.utils import ast_to_source_code
+from paddle.jit.dy2static.utils import is_paddle_api
 from .base_transformer import (
     BaseTransformer,
 )
