+++ python/paddle/jit/dy2static/early_return_transformer.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from paddle.utils import gast
-from paddle.fluid.dygraph.dygraph_to_static.static_analysis import (
+from paddle.jit.dy2static.static_analysis import (
     AstNodeWrapper,
 )
 from .base_transformer import (
