+++ python/paddle/jit/dy2static/error.py
@@ -17,17 +17,17 @@ import sys
 import traceback
 import linecache
 import re
-import numpy as np
 
-from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
+from .origin_info import (
     Location,
     OriginInfo,
     global_origin_info_map,
 )
-from paddle.fluid.dygraph.dygraph_to_static.utils import (
-    _is_api_in_module_helper,
+from .utils import (
     RE_PYMODULE,
 )
+
+__all__ = []
 
 ERROR_DATA = "Error data about original source code information and traceback."
