+++ python/paddle/jit/dy2static/logical_transformer.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from paddle.utils import gast
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
+from paddle.jit.dy2static.utils import ast_to_source_code
 from .base_transformer import (
     BaseTransformer,
 )
