+++ python/paddle/jit/dy2static/program_translator.py
@@ -26,38 +26,34 @@ from paddle.fluid.data_feeder import check_type
 from paddle.fluid.layers.utils import flatten
 from paddle.fluid.dygraph.base import param_guard
 from paddle.fluid.dygraph.base import switch_to_static_graph
-from paddle.fluid.dygraph.dygraph_to_static import error
-from paddle.fluid.dygraph.dygraph_to_static import logging_utils
-from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
+from . import error
+from . import logging_utils
+from .origin_info import (
     attach_origin_info,
-)
-from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
     create_and_update_origin_info_map,
-)
-from paddle.fluid.dygraph.dygraph_to_static.origin_info import (
     update_op_callstack_with_origin_info,
 )
+
 from .partial_program import (
     partial_program_from,
 )
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_func
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
-from paddle.fluid.dygraph.dygraph_to_static.utils import func_to_source_code
-from paddle.fluid.dygraph.dygraph_to_static.utils import input_specs_compatible
-from paddle.fluid.dygraph.dygraph_to_static.utils import type_name
-from paddle.fluid.dygraph.dygraph_to_static.utils import unwrap
-from paddle.fluid.dygraph.dygraph_to_static.utils import (
+from .utils import (
+    ast_to_func,
+    ast_to_source_code,
+    func_to_source_code,
+    input_specs_compatible,
+    type_name,
+    unwrap,
     make_hashable,
     ALREADY_D2S,
 )
-from paddle.fluid.dygraph.dygraph_to_static.function_spec import (
+from .function_spec import (
     FunctionSpec,
     _hash_spec_names,
-)
-from paddle.fluid.dygraph.dygraph_to_static.function_spec import (
     get_buffers,
     get_parameters,
 )
+
 
 from .ast_transformer import DygraphToStaticAst
