+++ python/paddle/jit/dy2static/tensor_shape_transformer.py
@@ -14,8 +14,8 @@
 
 from paddle.utils import gast
 
-from paddle.fluid.dygraph.dygraph_to_static.utils import ast_to_source_code
-from paddle.fluid.dygraph.dygraph_to_static.static_analysis import (
+from paddle.jit.dy2static.utils import ast_to_source_code
+from paddle.jit.dy2static.static_analysis import (
     AstNodeWrapper,
 )
 from .base_transformer import (
